requirements.txt
uvicorn
fastapi
pymongo
typing
pydantic
python-dotenv